v13.1.19

+++ accounts/models.py
@@ -7,7 +7,7 @@ class Profile(models.Model) :
     leagues=models.JSONField(default={"leagues":[]})
     squad=models.JSONField(default={"squad":[]})
     squad_name=models.CharField(null=True,blank=True)
-    current_balance=models.IntegerField(default=12)
+    current_balance=models.FloatField(default=12)
     wildcard=models.BooleanField(default=True)
     freehit=models.BooleanField(default=True)
     tripplecaptin=models.BooleanField(default=True)
